ft(add:restrict deleting of user)

+++ src/server/controllers/user.js
@@ -1,5 +1,6 @@
 const User = require('../models/userModel');
 const bcrypt = require('bcrypt');
+const jwt = require('jsonwebtoken');
 const { signToken } = require('../helpers/auth');
 
 module.exports = {
@@ -94,6 +95,10 @@ module.exports = {
 		}
 	},
 	deleteUser     : async (req, res) => {
+		//check for current user
+		const currentUser = await req.headers.authorization
+		const tokenArray = jwt.decode(currentUser.split(' ')[1])
+		if(!tokenArray.canEdit) return res.status(400).send("Request for Permission to Delete")
 		//delete User
 		try {
 			const user = await User.findOneAndDelete({ _id: req.params.userId });
